+++ R/head-to-head-helpers.R
@@ -76,7 +76,7 @@ print.h2h <- function(x, ...) {
 #' @export
 get_cr_players <- function(cr_data, players = NULL, ...) {
   if (is.null(players)) {
-    players <- sort(unique(to_longcr(cr_data, repair = TRUE)$player),
+    players <- sort(unique(as_longcr(cr_data, repair = TRUE)$player),
                     na.last = FALSE)
   }
 
@@ -87,7 +87,7 @@ get_cr_players <- function(cr_data, players = NULL, ...) {
 #' @export
 get_cr_matchups <- function(cr_data) {
   cr <- cr_data %>%
-    to_longcr(repair = TRUE) %>%
+    as_longcr(repair = TRUE) %>%
     select(.data$game, .data$player, .data$score)
 
   left_join(x = cr, y = cr, by = "game", suffix = c("1", "2"))